# openspec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PPSC MCQ Power Bank",
    "version": "0.0.1"
  },
  "paths": {
...
    "/categories/": {
      "post": {
        "tags": [
          "Categories",
          "categories"
        ],
        "summary": "Create Category",
        "description": "Create a new category with auto-generated slug if not provided.",
        "operationId": "create_category_categories__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Categories",
          "categories"
        ],
        "summary": "Get All Categories",
        "description": "Get all categories. By default, returns root categories. \nUse `include_subcategories=True` to also return children hierarchically.",
        "operationId": "get_all_categories_categories__get",
        "parameters": [
          {
            "name": "include_subcategories",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include nested subcategories",
              "default": false,
              "title": "Include Subcategories"
            },
            "description": "Include nested subcategories"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse_CategoryDetailResponse_"
                    },
                    {
                      "$ref": "#/components/schemas/PaginatedResponse_CategoryResponse_"
                    }
                  ],
                  "title": "Response Get All Categories Categories  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{slug}/subcategories": {
      "get": {
        "tags": [
          "Categories",
          "categories"
        ],
        "summary": "Get Subcategories",
        "description": "Get all subcategories for a given root category slug, with pagination.\nAvoids injecting empty 'subcategories' arrays into the output payload.",
        "operationId": "get_subcategories_categories__slug__subcategories_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_CategoryResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/categories/slug/{slug}": {
      "get": {
        "tags": [
          "Categories",
          "categories"
        ],
        "summary": "Get Category By Slug",
        "description": "Get a simple category by slug without hierarchy metadata.",
        "operationId": "get_category_by_slug_categories_slug__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{slug}/with-mcqs/{mcq_id}": {
      "get": {
        "tags": [
          "Categories",
          "categories"
        ],
        "summary": "Get Category Single Mcq",
        "description": "Get a single MCQ from a category with optional filtering.\n\nQuery Parameters:\n- explanation: If True, include/focus on explanation\n- with_mcq: If False with explanation=True, return only ID and explanation\n\nExamples:\n- /{slug}/with-mcqs/{id} - Full MCQ details (default)\n- /{slug}/with-mcqs/{id}?explanation=True - Full MCQ with explanation\n- /{slug}/with-mcqs/{id}?explanation=True&with_mcq=False - Only ID and explanation",
        "operationId": "get_category_single_mcq_categories__slug__with_mcqs__mcq_id__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "mcq_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mcq Id"
            }
          },
          {
            "name": "explanation",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Explanation"
            }
          },
          {
            "name": "with_mcq",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "With Mcq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{slug}/with-mcqs": {
      "get": {
        "tags": [
          "Categories",
          "categories"
        ],
        "summary": "Get Category With Mcqs",
        "description": "Get a category with its MCQs, with optional filtering and pagination.\n\nQuery Parameters:\n- explanation: If True, include/focus on explanations\n- with_mcq: If False with explanation=True, return only ID and explanation for each MCQ\n- include_subcategories: If True (default), include MCQs from all nested subcategories recursively.\n- limit: Maximum number of MCQs to return (default: 10, max: 100)\n- offset: Number of MCQs to skip for pagination (default: 0)\n\nExamples:\n- /{slug}/with-mcqs - First 10 MCQs with full details (default)\n- /{slug}/with-mcqs?limit=20 - First 20 MCQs\n- /{slug}/with-mcqs?limit=10&offset=10 - MCQs 11-20\n- /{slug}/with-mcqs?explanation=True - Full MCQs (explanations included in model)\n- /{slug}/with-mcqs?explanation=True&with_mcq=False - Only MCQ IDs and explanations",
        "operationId": "get_category_with_mcqs_categories__slug__with_mcqs_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "explanation",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Explanation"
            }
          },
          {
            "name": "with_mcq",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "With Mcq"
            }
          },
          {
            "name": "include_subcategories",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include MCQs from nested subcategories",
              "default": true,
              "title": "Include Subcategories"
            },
            "description": "Include MCQs from nested subcategories"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of MCQs to return",
              "default": 10,
              "title": "Limit"
            },
            "description": "Number of MCQs to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of MCQs to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of MCQs to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_CategoryDetailResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/categories/validate-slug/{slug}": {
      "get": {
        "tags": [
          "Categories",
          "categories"
        ],
        "summary": "Validate Slug",
        "description": "Check if a slug is valid (exists in database).",
        "operationId": "validate_slug_categories_validate_slug__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "human\u2010readable, e.g. 'PPSC All MCQs 2025'"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug",
            "description": "machine\u2010safe key (auto-generated if not provided)"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CategoryCreate"
      },
...
      "CategoryResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "title": "CategoryResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://ppsc-paper-bank.onrender.com",
      "description": "Render Production"
    }
  ]
}